+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/games/Game.java
@@ -436,15 +436,6 @@ public class Game {
     return null;
   }
 
-  @Nullable
-  @JsonIgnore
-  public File getPupPackFolder() {
-    if (!StringUtils.isEmpty(this.getRom())) {
-      return new File(new File(systemService.getPinUPSystemFolder(), "PUPVideos"), this.getRom());
-    }
-    return null;
-  }
-
 
   @Nullable
   @JsonIgnore