+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/puppack/PupPackUtil.java
@@ -3,6 +3,7 @@ package de.mephisto.vpin.server.puppack;
 import de.mephisto.vpin.restclient.jobs.JobExecutionResult;
 import de.mephisto.vpin.restclient.jobs.JobExecutionResultFactory;
 import edu.umd.cs.findbugs.annotations.NonNull;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -16,7 +17,7 @@ import java.util.zip.ZipInputStream;
 public class PupPackUtil {
   private final static Logger LOG = LoggerFactory.getLogger(PupPackUtil.class);
 
-  public static JobExecutionResult unzip(File archiveFile, File destinationDir, String rom) {
+  public static JobExecutionResult unzip(File archiveFile, File destinationDir, String rom, String tableName) {
     try {
       byte[] buffer = new byte[1024];
       FileInputStream fileInputStream = new FileInputStream(archiveFile);
@@ -25,13 +26,13 @@ public class PupPackUtil {
 
       while (zipEntry != null) {
         String name = zipEntry.getName();
-        File newFile = new File(destinationDir.getParentFile(), toTargetName(name, rom));
-        boolean isInPupPack = name.contains(rom + "/");
+        File newFile = new File(destinationDir, toTargetName(name, rom, tableName));
+        boolean isInPupPack = name.contains(rom + "/") || (!StringUtils.isEmpty(tableName) && name.contains(tableName));
         if (!isInPupPack) {
           LOG.info("Skipping extraction of " + newFile.getAbsolutePath());
         }
         else if (zipEntry.isDirectory()) {
-          if (!newFile.isDirectory() && !newFile.mkdirs()) {
+          if (!newFile.exists() && !newFile.mkdirs()) {
             throw new IOException("Failed to create directory " + newFile);
           }
         }
@@ -63,10 +64,19 @@ public class PupPackUtil {
   }
 
   @NonNull
-  private static String toTargetName(String name, String rom) {
-    while (!name.startsWith(rom + "/") && name.contains("/")) {
-      name = name.substring(name.indexOf("/") + 1);
+  private static String toTargetName(String name, String rom, String tableName) {
+    String targetFolder = name;
+    while (!targetFolder.startsWith(rom + "/") && targetFolder.contains("/")) {
+      targetFolder = targetFolder.substring(targetFolder.indexOf("/") + 1);
     }
-    return name;
+
+    if (!targetFolder.startsWith(rom) && !StringUtils.isEmpty(tableName)) {
+      targetFolder = name;
+      while (!targetFolder.startsWith(tableName + "/") && targetFolder.contains("/")) {
+        targetFolder = targetFolder.substring(targetFolder.indexOf("/") + 1);
+      }
+    }
+
+    return targetFolder;
   }
 }